+++ hydra-java/src/main/java/hydra/lib/literals/BigfloatToBigint.java
@@ -42,7 +42,7 @@ public class BigfloatToBigint extends PrimitiveFunction {
      * @param value the value
      * @return the result
      */
-        public static BigInteger apply(Double value) {
-        return BigDecimal.valueOf(value).toBigInteger();
+        public static BigInteger apply(BigDecimal value) {
+        return value.toBigInteger();
     }
 }
